--- lib/__init__.py
@@ -1,4 +1,0 @@
-# pylint: disable=missing-module-docstring
-from . import models
-from . import scenarios
-from . import post_processing
