--- setup.py
@@ -1,8 +1,0 @@
-# pylint: disable=missing-module-docstring
-from setuptools import setup, find_packages
-
-setup(
-    name="lib",
-    version="0.1",
-    packages=find_packages(),
-)
